v1.0.7 -- fix read_counts() to add support for very large integers (> 32bit signed integers such as 2447935369)

+++ R/read_counts.R
@@ -107,7 +107,7 @@ read_counts <- function(counts_file, samples = NULL) {
         data.table::fread(
             counts_file,
             skip = 2,
-            colClasses = c("character", rep("integer", ncol(counts_info) - 1)),
+            colClasses = c("character", rep("numeric", ncol(counts_info) - 1)),
             nThread = 1,
             sep = "\t",
             data.table = FALSE,
